backend/pyproject.toml
[project]
name = "skai-backend"
version = "0.1.0"
description = "FastAPI backend for SKAI — AI-powered restaurant phone ordering system."
requires-python = ">=3.12,<3.14"
dependencies = [
  "elevenlabs>=1.0",
  "fastapi>=0.115.0",
  "google-genai",
  "httpx>=0.27",
  "pydantic-settings>=2.0",
  "pydantic>=2.0",
  "python-dateutil>=2.9",
  "python-jose[cryptography]>=3.3",
  "python-multipart>=0.0.9",
  "supabase>=2.0",
  "twilio>=9.0",
  "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = []

